Update user (#3)
* добавить enum Gender;

* добавить поля в NewUserRequest;

* добавить поля в User;

* добавить поля в UserDto;

* заменить список ссылок соцсетей на одну соцсеть;
