+++ src/main/java/com/kardoaward/kardo/user/mapper/UserMapper.java
@@ -1,8 +1,8 @@
 package com.kardoaward.kardo.user.mapper;
 
-import com.kardoaward.kardo.user.model.NewUserRequest;
+import com.kardoaward.kardo.user.model.dto.NewUserRequest;
 import com.kardoaward.kardo.user.model.User;
-import com.kardoaward.kardo.user.model.UserDto;
+import com.kardoaward.kardo.user.model.dto.UserDto;
 import org.mapstruct.Mapper;
 
 @Mapper(componentModel = "spring")
