+++ src/main/java/com/kardoaward/kardo/user/model/dto/UserDto.java
@@ -1,34 +1,17 @@
-package com.kardoaward.kardo.user.model;
+package com.kardoaward.kardo.user.model.dto;
 
 import com.kardoaward.kardo.user.model.enums.Gender;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import jakarta.persistence.Transient;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
+import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
 
 import java.time.LocalDate;
 
-@Entity
-@Table(name = "users")
+@Data
 @AllArgsConstructor
 @NoArgsConstructor
-@Getter
-@Setter
-@ToString
-public class User {
+public class UserDto {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String email;
     private String name;
@@ -38,17 +21,12 @@ public class User {
     private String country;
     private String region;
     private String city;
-    @Column(name = "phone_number")
     private String phoneNumber;
-    @Enumerated(EnumType.STRING)
     private Gender gender;
     private String citizenship;
-    @Column(name = "profile_photo")
     private String profilePhoto;
     private String portfolio;
-    @Column(name = "about_me")
     private String aboutMe;
     //ToDo Исправить на список. В БД это поле отсутствует.
-    @Transient
     private String network;
 }
