+++ src/main/java/com/kardoaward/kardo/user/model/enums/Gender.java
@@ -1,0 +1,6 @@
+package com.kardoaward.kardo.user.model.enums;
+
+public enum Gender {
+    MALE,
+    FEMALE
+}
